requirements.txt
aiogram==3.3.0
pytesseract
pillow
aiohttp
